+++ src/main/java/com/witcurve/service/util/PayrollComparator.java
@@ -7,7 +7,7 @@ import java.util.Comparator;
 public class PayrollComparator implements Comparator<PayrollDTO> {
     @Override
     public int compare(PayrollDTO o1, PayrollDTO o2) {
-        if (o1.getPayrollCycle().getYear() == o2.getPayrollCycle().getYear()) {
+        if (o1.getPayrollCycle().getYear().intValue() == o2.getPayrollCycle().getYear().intValue()) {
             if (o1.getPayrollCycle().getMonth().getValue() == o2.getPayrollCycle().getMonth().getValue()) {
                 return 0;
             } else if (o1.getPayrollCycle().getMonth().getValue() > o2.getPayrollCycle().getMonth().getValue()) {
